Revert "REF more module level changes to code"
This reverts commit 81657807b51b036514177af02a7f6a6012101e74.
