+++ ngmix/priors/__init__.py
@@ -1,5 +1,3 @@
 # flake8: noqa
 from ._priors import *
 from ._random import make_rng, srandu
-from ._constants import LOWVAL, BIGVAL
-from ._base import PriorBase, GPriorBase
