--- ngmix/priors/_constants.py
@@ -1,4 +1,0 @@
-import numpy
-
-LOWVAL = -numpy.inf
-BIGVAL = 9999.0e47
